requirements.txt
aiopg==1.0.0
alabaster==0.7.12
attrs==19.3.0
Babel==2.7.0
bcrypt==3.1.7
cachetools==4.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Deprecated==1.2.7
docutils==0.15.2
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.8
imagesize==1.1.0
importlib-metadata==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
parsimonious==0.7.0
pathlib2==2.3.5
pluggy==0.13.1
protobuf==3.14.0
psycopg2==2.8.4
psycopg2-binary==2.8.5
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.44.1
Pygments==2.5.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.5
pytest==5.3.1
pytest-tornado==0.8.0
pytz==2019.3
requests==2.22.0
rsa==4.6
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.2
sphinx-js==2.8
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
